move CovariancePC to top level namespace
