+++ firedrake/adjoint/__init__.py
@@ -35,9 +35,8 @@ from firedrake.adjoint.ufl_constraints import (  # noqa: F401
 from firedrake.adjoint.ensemble_reduced_functional import EnsembleReducedFunctional  # noqa F401
 from firedrake.adjoint.transformed_functional import L2RieszMap, L2TransformedFunctional  # noqa: F401
 from firedrake.adjoint.covariance_operator import (  # noqa F401
-    WhiteNoiseGenerator, AutoregressiveCovariance,
-    PyOP2NoiseBackend, PetscNoiseBackend, VOMNoiseBackend,
-    CovarianceMat, CovariancePC)
+    WhiteNoiseGenerator, AutoregressiveCovariance, CovarianceMat,
+    PyOP2NoiseBackend, PetscNoiseBackend, VOMNoiseBackend)
 import numpy_adjoint  # noqa F401
 import firedrake.ufl_expr
 import types
